Student Access & differentiation of roles

+++ app/Http/Controllers/AbstractCrudController.php
@@ -138,19 +138,24 @@ class AbstractCrudController extends Controller implements IAbstractCrudControll
         }
     }
 
-    public static function setupRouter(Router $router)
+    public static function setupRouter(Router $router, array $functionality = ['findAll', 'findById', 'create', 'update', 'delete'])
     {
         try {
             $className = (new \ReflectionClass(static::class))->getShortName();
         } catch (\ReflectionException $exception) {
             throw new \RuntimeException("Can't setupRouter on \"" . static::class . "\"", 500);
         }
-        $router->group(["prefix" => camel_case($className)], function (Router $router) use ($className) {
-            $router->get(null, $className . "@findAll");
-            $router->get("{id}", $className . "@findById");
-            $router->post(null, $className . "@create");
-            $router->put("{id}", $className . "@update");
-            $router->delete("{id}", $className . "@delete");
+        $router->group(["prefix" => camel_case($className)], function (Router $router) use ($className, $functionality) {
+            if (in_array('findAll', $functionality))
+                $router->get(null, $className . "@findAll");
+            if (in_array('findById', $functionality))
+                $router->get("{id}", $className . "@findById");
+            if (in_array('create', $functionality))
+                $router->post(null, $className . "@create");
+            if (in_array('update', $functionality))
+                $router->put("{id}", $className . "@update");
+            if (in_array('delete', $functionality))
+                $router->delete("{id}", $className . "@delete");
         });
     }
